tools: move ESLint to tools/eslint
Greatly simplify how ESLint and its plugins are installed.
