+++ test/common/index.js
@@ -499,7 +499,7 @@ function hasMultiLocalhost() {
 
 function skipIfEslintMissing() {
   if (!fs.existsSync(
-    path.join(__dirname, '..', '..', 'tools', 'node_modules', 'eslint'),
+    path.join(__dirname, '..', '..', 'tools', 'eslint', 'node_modules', 'eslint'),
   )) {
     skip('missing ESLint');
   }
