+++ test/parallel/test-eslint-alphabetize-errors.js
@@ -6,7 +6,7 @@ if ((!common.hasCrypto) || (!common.hasIntl)) {
 }
 common.skipIfEslintMissing();
 
-const RuleTester = require('../../tools/node_modules/eslint').RuleTester;
+const RuleTester = require('../../tools/eslint/node_modules/eslint').RuleTester;
 const rule = require('../../tools/eslint-rules/alphabetize-errors');
 
 new RuleTester().run('alphabetize-errors', rule, {
