+++ test/parallel/test-eslint-async-iife-no-unused-result.js
@@ -5,7 +5,7 @@ if ((!common.hasCrypto) || (!common.hasIntl)) {
 }
 common.skipIfEslintMissing();
 
-const RuleTester = require('../../tools/node_modules/eslint').RuleTester;
+const RuleTester = require('../../tools/eslint/node_modules/eslint').RuleTester;
 const rule = require('../../tools/eslint-rules/async-iife-no-unused-result');
 
 const message = 'The result of an immediately-invoked async function needs ' +
