+++ test/parallel/test-eslint-duplicate-requires.js
@@ -7,7 +7,7 @@ if ((!common.hasCrypto) || (!common.hasIntl)) {
 
 common.skipIfEslintMissing();
 
-const { RuleTester } = require('../../tools/node_modules/eslint');
+const { RuleTester } = require('../../tools/eslint/node_modules/eslint');
 const rule = require('../../tools/eslint-rules/no-duplicate-requires');
 
 new RuleTester({
