+++ test/parallel/test-eslint-lowercase-name-for-primitive.js
@@ -7,7 +7,7 @@ if ((!common.hasCrypto) || (!common.hasIntl)) {
 
 common.skipIfEslintMissing();
 
-const RuleTester = require('../../tools/node_modules/eslint').RuleTester;
+const RuleTester = require('../../tools/eslint/node_modules/eslint').RuleTester;
 const rule = require('../../tools/eslint-rules/lowercase-name-for-primitive');
 
 new RuleTester().run('lowercase-name-for-primitive', rule, {
