+++ test/parallel/test-eslint-no-array-destructuring.js
@@ -7,7 +7,7 @@ if ((!common.hasCrypto) || (!common.hasIntl)) {
 
 common.skipIfEslintMissing();
 
-const { RuleTester } = require('../../tools/node_modules/eslint');
+const { RuleTester } = require('../../tools/eslint/node_modules/eslint');
 const rule = require('../../tools/eslint-rules/no-array-destructuring');
 
 const USE_OBJ_DESTRUCTURING =
