+++ test/parallel/test-eslint-no-unescaped-regexp-dot.js
@@ -7,7 +7,7 @@ if ((!common.hasCrypto) || (!common.hasIntl)) {
 
 common.skipIfEslintMissing();
 
-const RuleTester = require('../../tools/node_modules/eslint').RuleTester;
+const RuleTester = require('../../tools/eslint/node_modules/eslint').RuleTester;
 const rule = require('../../tools/eslint-rules/no-unescaped-regexp-dot');
 
 new RuleTester().run('no-unescaped-regexp-dot', rule, {
