+++ test/parallel/test-eslint-prefer-common-mustnotcall.js
@@ -7,7 +7,7 @@ if ((!common.hasCrypto) || (!common.hasIntl)) {
 
 common.skipIfEslintMissing();
 
-const RuleTester = require('../../tools/node_modules/eslint').RuleTester;
+const RuleTester = require('../../tools/eslint/node_modules/eslint').RuleTester;
 const rule = require('../../tools/eslint-rules/prefer-common-mustnotcall');
 
 const message = 'Please use common.mustNotCall(msg) instead of ' +
