+++ test/parallel/test-eslint-prefer-common-mustsucceed.js
@@ -7,7 +7,7 @@ if ((!common.hasCrypto) || (!common.hasIntl)) {
 
 common.skipIfEslintMissing();
 
-const RuleTester = require('../../tools/node_modules/eslint').RuleTester;
+const RuleTester = require('../../tools/eslint/node_modules/eslint').RuleTester;
 const rule = require('../../tools/eslint-rules/prefer-common-mustsucceed');
 
 const msg1 = 'Please use common.mustSucceed instead of ' +
