+++ tools/eslint/eslint.config_utils.mjs
@@ -1,6 +1,6 @@
 import { createRequire } from 'node:module';
 
-export const requireEslintTool = createRequire(new URL('../node_modules/eslint/', import.meta.url));
+export const requireEslintTool = createRequire(new URL(import.meta.url));
 export const resolveEslintTool = (request) => requireEslintTool.resolve(request);
 
 export const noRestrictedSyntaxCommonAll = [
