+++ tools/license-builder.sh
@@ -107,8 +107,8 @@ licenseText="$(sed -e '/^$/,$d' -e 's/^#$//' -e 's/^# //' "${rootdir}/tools/cppl
 addlicense "cpplint.py" "tools/cpplint.py" "$licenseText"
 licenseText="$(sed -e '/^$/,$d' -e 's/^#$//' -e 's/^# //' "${rootdir}/tools/gypi_to_gn.py" | tail -n +3)"
 addlicense "gypi_to_gn.py" "tools/gypi_to_gn.py" "$licenseText"
-licenseText="$(cat "${rootdir}/tools/node_modules/eslint/LICENSE")"
-addlicense "ESLint" "tools/node_modules/eslint" "$licenseText"
+licenseText="$(cat "${rootdir}/tools/eslint/node_modules/eslint/LICENSE")"
+addlicense "ESLint" "tools/eslint/node_modules/eslint" "$licenseText"
 licenseText="$(cat "${rootdir}/deps/googletest/LICENSE")"
 addlicense "gtest" "deps/googletest" "$licenseText"
